Use device dependent intrinsic parameters and make window size flexible
As the depth images were having too much of flying pixels, I though of using the intrinsic
parameters of the device I have used -- That did fix the problem. I also fixed the window
size that is required for getting rid of flying pixels. Now, the window size could be
defined by the user instead of having just a size of 2.